Update cpdb_statistical_analysis_helper.py
speed up the `build_clusters` step which appears to be a bottleneck.

Instead of using pandas's slicing method, convert the original df to numpy ndarray and slice using the column position with a list comprehension and then use np.mean to calculate the columns means quickly.

The output is changed back into a pd.Series to fit the original structure. The complex_composition step seems to just query and retrieve the `clusters['means']` and `clusters['counts']` from above, so shouldn't take too long? need to test...